+++ xbmc/addons/AddonBindings.cmake
@@ -6,7 +6,7 @@
 set(CORE_ADDON_BINDINGS_FILES
     ${CORE_SOURCE_DIR}/xbmc/cores/AudioEngine/Utils/AEChannelData.h
     ${CORE_SOURCE_DIR}/xbmc/filesystem/IFileTypes.h
-    ${CORE_SOURCE_DIR}/xbmc/input/ActionIDs.h
+    ${CORE_SOURCE_DIR}/xbmc/input/actions/ActionIDs.h
     ${CORE_SOURCE_DIR}/xbmc/input/XBMC_vkeys.h
 )
