+++ xbmc/dialogs/GUIDialogTextViewer.cpp
@@ -11,8 +11,8 @@
 #include "filesystem/File.h"
 #include "guilib/GUIComponent.h"
 #include "guilib/GUIWindowManager.h"
-#include "input/Action.h"
-#include "input/ActionIDs.h"
+#include "input/actions/Action.h"
+#include "input/actions/ActionIDs.h"
 #include "utils/log.h"
 #include "utils/URIUtils.h"
 #include "ServiceBroker.h"
