+++ xbmc/games/dialogs/osd/DialogGameVideoSelect.cpp
@@ -14,7 +14,7 @@
 #include "guilib/GUIMessage.h"
 #include "guilib/GUIComponent.h"
 #include "guilib/GUIWindowManager.h"
-#include "input/ActionIDs.h"
+#include "input/actions/ActionIDs.h"
 #include "settings/GameSettings.h"
 #include "settings/MediaSettings.h"
 #include "settings/Settings.h"
