+++ xbmc/input/CMakeLists.txt
@@ -1,6 +1,4 @@
-set(SOURCES Action.cpp
-            ActionTranslator.cpp
-            AppTranslator.cpp
+set(SOURCES AppTranslator.cpp
             ButtonTranslator.cpp
             CustomControllerTranslator.cpp
             GamepadTranslator.cpp
@@ -27,9 +25,6 @@ set(SOURCES Action.cpp
 
 set(HEADERS hardware/IHardwareInput.h
             remote/IRRemote.h
-            Action.h
-            ActionIDs.h
-            ActionTranslator.h
             AppTranslator.h
             ButtonTranslator.h
             CustomControllerTranslator.h
