+++ xbmc/input/InputManager.h
@@ -13,8 +13,9 @@
 #include <string>
 #include <vector>
 
-#include "Action.h"
 #include "windowing/XBMC_events.h"
+#include "input/actions/Action.h"
+#include "input/mouse/interfaces/IMouseInputProvider.h"
 #include "input/mouse/MouseStat.h"
 #include "input/KeyboardStat.h"
 #include "interfaces/IActionListener.h"
