+++ xbmc/input/JoystickMapper.cpp
@@ -7,8 +7,8 @@
  */
 
 #include "JoystickMapper.h"
-#include "ActionIDs.h"
-#include "ActionTranslator.h"
+#include "input/actions/ActionIDs.h"
+#include "input/actions/ActionTranslator.h"
 #include "input/joysticks/JoystickTranslator.h"
 #include "input/joysticks/JoystickUtils.h"
 #include "input/WindowKeymap.h"
