+++ xbmc/input/Key.h
@@ -14,8 +14,8 @@
  */
 
 //! @todo Remove dependence on CAction
-#include "Action.h"
-#include "ActionIDs.h"
+#include "input/actions/Action.h"
+#include "input/actions/ActionIDs.h"
 
 #include <string>
 #include <stdint.h>
