+++ xbmc/input/actions/ActionTranslator.cpp
