+++ xbmc/input/actions/CMakeLists.txt
@@ -1,0 +1,10 @@
+set(SOURCES Action.cpp
+            ActionTranslator.cpp
+)
+
+set(HEADERS Action.h
+            ActionIDs.h
+            ActionTranslator.h
+)
+
+core_add_library(input_actions)
