+++ xbmc/input/joysticks/keymaps/KeyHandler.h
@@ -8,9 +8,9 @@
 
 #pragma once
 
+#include "input/actions/Action.h"
 #include "input/joysticks/interfaces/IKeyHandler.h"
 #include "input/joysticks/JoystickTypes.h"
-#include "input/Action.h"
 
 #include <map>
 #include <string>
