+++ xbmc/input/keyboard/KeymapActionMap.cpp
@@ -9,7 +9,7 @@
 #include "KeymapActionMap.h"
 #include "guilib/GUIComponent.h"
 #include "guilib/GUIWindowManager.h"
-#include "input/Action.h"
+#include "input/actions/Action.h"
 #include "input/InputManager.h"
 #include "input/Key.h"
 #include "ServiceBroker.h"
