+++ xbmc/interfaces/swig/AddonModuleXbmcgui.i
@@ -20,7 +20,7 @@
 #include "interfaces/legacy/WindowDialog.h"
 #include "interfaces/legacy/Dialog.h"
 #include "interfaces/legacy/WindowXML.h"
-#include "input/ActionIDs.h"
+#include "input/actions/ActionIDs.h"
 #include "input/Key.h"
 
 using namespace XBMCAddon;
@@ -109,5 +109,5 @@ using namespace xbmcgui;
 
 %include "interfaces/legacy/WindowXML.h"
 
-%include "input/ActionIDs.h"
+%include "input/actions/ActionIDs.h"
 %include "input/Key.h"
