+++ xbmc/network/EventServer.cpp
@@ -7,22 +7,24 @@
  */
 
 #include "EventServer.h"
-#include "EventPacket.h"
 #include "EventClient.h"
+#include "EventPacket.h"
 #include "Socket.h"
-#include "Application.h"
-#include "ServiceBroker.h"
-#include "interfaces/builtins/Builtins.h"
-#include "input/ActionTranslator.h"
-#include "threads/SingleLock.h"
 #include "Zeroconf.h"
 #include "guilib/GUIAudioManager.h"
+#include "input/actions/ActionTranslator.h"
 #include "input/Key.h"
+#include "interfaces/builtins/Builtins.h"
+#include "threads/SingleLock.h"
 #include "utils/log.h"
 #include "utils/SystemInfo.h"
+#include "Application.h"
+#include "ServiceBroker.h"
+#include "Util.h"
+
+#include <cassert>
 #include <map>
 #include <queue>
-#include <cassert>
 
 using namespace EVENTSERVER;
 using namespace EVENTPACKET;
