+++ xbmc/pvr/dialogs/GUIDialogPVRItemsViewBase.cpp
@@ -12,8 +12,8 @@
 #include "FileItem.h"
 #include "ServiceBroker.h"
 #include "dialogs/GUIDialogContextMenu.h"
-#include "input/Action.h"
-#include "input/ActionIDs.h"
+#include "input/actions/Action.h"
+#include "input/actions/ActionIDs.h"
 #include "view/ViewState.h"
 
 #include "pvr/PVRGUIActions.h"
