+++ xbmc/video/dialogs/GUIDialogVideoOSD.cpp
@@ -12,7 +12,7 @@
 #include "guilib/GUIComponent.h"
 #include "guilib/GUIWindowManager.h"
 #include "guilib/WindowIDs.h"
-#include "input/ActionIDs.h"
+#include "input/actions/ActionIDs.h"
 #include "input/InputManager.h"
 #include "ServiceBroker.h"
